% Generated by roxygen2 (4.0.0): do not edit by hand
\name{datas transformation}
\alias{datas transformation}
\alias{date2month}
\alias{date2summer}
\alias{date2wy}
\alias{date2year}
\title{Functions to transform dates into other useful information}
\usage{
date2year(datevec)

date2wy(datevec, endMonth = 9)

date2month(datevec)

date2summer(datevec, begMonth = 4, endMonth = 9)
}
\arguments{
\item{datevec}{a vetor of dates (\code{class(x)} must be Date)}
}
\value{
The length of \code{x} without the NA
}
\description{
Functions to transform dates into other useful information
}

