% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesManip.R
\name{dates transformation}
\alias{dates transformation}
\alias{date2year}
\alias{date2wy}
\alias{date2month}
\alias{date2summer}
\title{Functions to transform dates into other useful information}
\usage{
date2year(datevec)

date2wy(datevec, endMonth = 9)

date2month(datevec)

date2summer(datevec, begMonth = 4, endMonth = 9)
}
\arguments{
\item{datevec}{a vetor of dates (\code{class(datevec)} must be Date)}

\item{endMonth}{the month which indicates the end of the Water Year or Summer Season (default is 9, September)}

\item{begMonth}{the month which indicates the beginning of summer  (default is 4, April)}
}
\value{
The year, water year or month from the input date. Or a 0-1 indicator of summer.
}
\description{
Functions to transform dates into other useful information
}
\concept{manipulate dates}
