% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Generalised Logistic distibution}
\alias{Generalised Logistic distibution}
\alias{dglo}
\alias{pglo}
\alias{qglo}
\alias{rglo}
\title{The Generalised Logistic distibution}
\usage{
dglo(x, loc, scale, sh)

pglo(q, loc, scale, sh, lower.tail = TRUE)

qglo(p, loc, scale, sh, lower.tail = TRUE)

rglo(n, loc, scale, sh)
}
\arguments{
\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{sh}{shape parameter}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{x,q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}}
}
\value{
These function mimic the standard output of distributions in R see \code{?pnorm}.
}
\description{
Density, distribution function, quantile function and random generation for the Generalised Logistic distribution (as in Hosking and Wallis' book) with location parameter equal to \code{loc}, scale paraemter equale to \code{scale} and shape parameter equal to \code{shape}
}
\examples{
plot(seq(-26,80,by=0.2),dglo(seq(-26,80,by=0.2),4,6,-0.2),type="l")
plot(ecdf(rglo(100,4,6,-0.2)))
lines(seq(-26,80,by=0.2),pglo(seq(-26,80,by=0.2),4,6,-0.2),col=2)
qglo(c(0.5,0.99,0.995,0.995,0.999),4,6,-0.2) # the 1-in-2years, 1-in-100years, 1-in-200years, 1-in-1000years events
}

