% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevglo.R
\name{Generalised Logistic distribution}
\alias{Generalised Logistic distribution}
\alias{dglo}
\alias{pglo}
\alias{qglo}
\alias{rglo}
\title{The Generalised Logistic distribution}
\usage{
dglo(x, loc, scale, sh, log = FALSE)

pglo(q, loc, scale, sh, lower.tail = TRUE, log.p = FALSE)

qglo(p, loc, scale, sh, lower.tail = TRUE, log.p = FALSE)

rglo(n, loc, scale, sh)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{sh}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}}

\item{p}{vector of probabilities}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
dglo gives the density, pglo gives the distribution function, qglo gives the quantile function, 
and rglo generates random deviates. The length of the result is determined by n for rglo, 
and is the maximum of the lengths of the numerical arguments for the other functions.
The numerical arguments are recycled to the length of the result. 
Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for the
Generalised Logistic distribution (as in Hosking and Wallis' book) with location parameter 
equal to \code{loc}, scale parameter equal to \code{scale} and shape parameter equal to \code{sh}
}
\examples{
plot(seq(-26,80,by=0.2),dglo(seq(-26,80,by=0.2),4,6,-0.2),type="l")
plot(ecdf(rglo(100,4,6,-0.2)))
lines(seq(-26,80,by=0.2),pglo(seq(-26,80,by=0.2),4,6,-0.2),col=2)
qglo(c(0.5,0.99,0.995,0.995,0.999),4,6,-0.2) 
# notable quantiles
}
\references{
Hosking, J.R.M. and Wallis, J.R., 2005. Regional frequency analysis: an approach based on L-moments. Cambridge university press.
}
\concept{glo distribution}
