% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/withHist.R
\name{gev.hist.fit}
\alias{gev.hist.fit}
\title{Maximum Likelihood Fitting for the GEV distibution in the presence of Historical Data}
\usage{
gev.hist.fit(xdat, k = 0, h = NULL, X0 = NULL, binomialcens = FALSE,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{xdat}{vector of historical and systematic/observed data - the first k elements of the vector should be the historical events}

\item{k}{number of historical events available. These events should be stored as the first observations of the xdat vector}

\item{h}{length of years covered by the information of the historical period}

\item{X0}{the perception threshold which is exceeded by the historical events. This is most likely different from the lowest historical value. It should indicate a value after which we are confident the event would have been recorded or left traces}

\item{binomialcens}{a logical value. Indicates whether the actual k values are to be used, or if only the information that the threshold X0 has been exceeded is used.}

\item{mulink}{the link function for the location parameter - default to identity}

\item{siglink}{the link function for the scale parameter - default to identity}

\item{shlink}{the link function for the shape parameter - default to identity}

\item{muinit}{initial values for the location parameter}

\item{siginit}{initial values for the scale parameter}

\item{shinit}{initial values for the shape parameter}

\item{method}{The optimization method (see \code{optim} for details)}

\item{maxit}{The maximum number of iterations}

\item{...}{Other control parameters for the optimization. These are passed to components of the control argument of optim.}
}
\value{
An object of the class \code{gev.fit}.
}
\description{
ML fitting procedures for sample of both systematic and historical data. Function structure inspired by \code{gev.fit} in ismev.
Arguments naming and likelihood implementation follow Stedinger and Cohn (1986).
}
\details{
The form of the GEV used is that of Coles (2001) Eq (3.2). Specifically, positive values of the shape parameter imply a heavy tail, and negative values imply a bounded upper tail.
}
\examples{
# library(ismev)
set.seed(5416574)
xx <- rgev(500, 40, 6, -0.2)
xxsist <- xx[471:500]; xxhist <- xx[1:470][xx[1:470] > 80]
gev.hist.fit(c(xxhist,xxsist), k = 3, h = 470, X0 = 80)
gev.hist.fit(c(xxhist,xxsist), k = 3, h = 470, X0 = 80, binomialcens = TRUE)
ismev::gev.fit(xxsist) # note the higher standard errors
}

