% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{gevcvd.fit}
\alias{gevcvd.fit}
\title{Maximum Likelihood Fitting for the GEV distribution - CV model}
\usage{
gevcvd.fit(
  xdat,
  ydat = NULL,
  mul = NULL,
  taul = NULL,
  shl = NULL,
  mulink = identity,
  taulink = identity,
  shlink = identity,
  muinit = NULL,
  tauinit = NULL,
  shinit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  optimPars = NULL,
  maxit = 10000,
  fixedPars = list(mu = NULL, tau = NULL, sh = NULL),
  ...
)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted}

\item{ydat}{A matrix of covariates for generalized linear modelling of the parameters (or NULL (the default) for stationary fitting). The number of rows should be the same as the length of xdat}

\item{mul}{Numeric vectors of integers, giving the columns of ydat that contain covariates for generalized linear modelling of the location parameter (or NULL (the default) if the corresponding parameter is stationary)}

\item{taul}{As \code{mul} for the tau parameter}

\item{shl}{As \code{mul} for the shape parameter}

\item{mulink}{the link function for the location parameter - default to identity}

\item{taulink}{the link function for the scale parameter - default to identity}

\item{shlink}{the link function for the shape parameter - default to identity}

\item{muinit}{initial values for the location parameter}

\item{tauinit}{initial values for the tau parameter}

\item{shinit}{initial values for the shape parameter}

\item{show}{Logical; if \code{TRUE} (the default), print details of the fit.}

\item{method}{The optimization method (see \code{optim} for details)}

\item{optimPars}{A string with other parameters to pass into \code{optim}. For example, depending on \code{method}, one could have "lower = 10, upper = 20"}

\item{maxit}{The maximum number of iterations}

\item{fixedPars}{a named list to fix any of the distribution parameter to a given value. When the named parameter is set to \code{NULL} its value is estimated.}

\item{...}{Other control parameters for the optimization. These are passed to components of the control argument of optim.}
}
\value{
An object of the \code{gevcd.fit} class, similar to \code{gev.fit} objects obtained using ismev::gev.fit. 
The \code{ismev} functions for gev.fit will not work on these objects. 

In the output the \code{vals} matrix gives the location and scale values obtained as scale = \eqn{\tau} *  location.
}
\description{
Maximum-likelihood fitting for the generalized extreme value distribution, 
including generalized linear modelling of each parameter. The function differs from 
\code{gevd.fit} because it uses a different parametrisation of the distribution based on 
the \eqn{\tau}, the ratio of the scale parameter and the location parameter, which is a monotonic 
function of the coefficient of variation. This means that when regression models are applied for the location,
these also affect the scale. 
The function allows any parameter to be kept fixed and to not be estimated.
}
\examples{
set.seed(12)
x <- runif(500)
y <- rgev(500,loc = 40+4*x, scale = 0.2*(40+4*x), sh = 0.15)
fit1 <- gevcvd.fit(y, show=FALSE)
fit1
## now add a regression model for the location
fit2 <- gevcvd.fit(y, ydat = cbind(x), mul=1, show=FALSE)
fit2
## now a fit with a fixed tau parameter
fitf <- gevcvd.fit(y, ydat = cbind(x), mul=1, show=FALSE, fixedPars = list(tau = 0.2))
fitf ## only two parameters are estimated (location and shape)
}
