% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{gevd.fit}
\alias{gevd.fit}
\title{Maximum Likelihood Fitting for the GEV distribution}
\usage{
gevd.fit(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", optimPars = NULL, maxit = 10000,
  fixedPars = list(mu = NULL, sig = NULL, sh = NULL), ...)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted}

\item{ydat}{A matrix of covariates for generalized linear modelling of the parameters (or NULL (the default) for stationary fitting). The number of rows should be the same as the length of xdat}

\item{mul}{Numeric vectors of integers, giving the columns of ydat that contain covariates for generalized linear modelling of the location parameter (or NULL (the default) if the corresponding parameter is stationary)}

\item{sigl}{As \code{mul} for the scale parameter}

\item{shl}{As \code{mul} for the shape parameter}

\item{mulink}{the link function for the location parameter - default to identity}

\item{siglink}{the link function for the scale parameter - default to identity}

\item{shlink}{the link function for the shape parameter - default to identity}

\item{muinit}{initial values for the location parameter}

\item{siginit}{initial values for the scale parameter}

\item{shinit}{initial values for the shape parameter}

\item{show}{Logical; if \code{TRUE} (the default), print details of the fit.}

\item{method}{The optimization method (see \code{optim} for details).}

\item{optimPars}{A string with other parameters to pass into \code{optim}. For example, depending on \code{method}, one could have "lower = 10, upper = 20"}

\item{maxit}{The maximum number of iterations}

\item{fixedPars}{a named list to fix any of the distribution parameter to a given value. When the named parameter is set to \code{NULL} its value is estimated.}

\item{...}{Other control parameters for the optimization. These are passed to components of the control argument of optim.}
}
\value{
An object of the gev.fit class as for objects obtained using ismev::gev.fit, so that \code{ismev} functions
for gev.fit work on these objects when no parameter is kept fixed (i.e. \code{fixedPars} is a list of NULL)
}
\description{
Maximum-likelihood fitting for the generalized extreme value distribution, 
including generalized linear modelling of each parameter. 
This function has the same structure as the \code{gev.fit} from \code{ismev}. 
Additionally it allows any parameter to be kept fixed and to not be estimated.
}
\details{
The form of the GEV used is that of Coles (2001) Eq (3.2) - so it is not the same as the one used in \link{dgev}. 
Specifically, positive values of the shape parameter imply a heavy tail, and negative values imply a bounded upper tail.
}
\examples{
set.seed(12)
x <- runif(500)
y <- rgev(500,loc = 40+4*x,scale = 6,sh = 0.2)
fit1 <- gevd.fit(y, show=FALSE)
fit1
## now add a regression model for the location
fit2 <- gevd.fit(y, ydat = cbind(x), mul=1, show=FALSE)
fit2
## now a fit with a fixed shape parameter (notice the sign)
fitf <- gevd.fit(y, show=FALSE, fixedPars = list(sh = -0.2))
fitf ## only two parameters are estimated 
}
\references{
Hosking, J.R.M. and Wallis, J.R., 2005. Regional frequency analysis: an approach based on L-moments. Cambridge university press. 

Coles, S., 2001. An introduction to statistical modeling of extreme values. London: Springer.
}
\seealso{
\link{dgev}
}
