% Generated by roxygen2 (4.0.0): do not edit by hand
\name{glo.fit}
\alias{glo.fit}
\title{Fitting the GLo distribution}
\usage{
glo.fit(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{xdat}{the object to be fitted}
}
\description{
This function fits the GLo distribution. It's built mirroring the \code{gev.fit} function of \code{ismev}
}
\examples{
library(ismev)
print(glo.fit(c(50,45,65,78,12,23),show=FALSE))
}
\keyword{glo.fit}

