% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withHist.R
\name{glo.hist.fit}
\alias{glo.hist.fit}
\title{Maximum Likelihood Fitting for the GLO distribution in the presence of Historical Data}
\usage{
glo.hist.fit(
  xdat,
  k = 0,
  h = NULL,
  X0 = NULL,
  binomialcens = FALSE,
  mulink = identity,
  siglink = identity,
  shlink = identity,
  muinit = NULL,
  siginit = NULL,
  shinit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 10000,
  ...
)
}
\arguments{
\item{xdat}{vector of historical and systematic/observed data - the first k elements of the vector should be the historical events}

\item{k}{number of historical events available. These events should be stored as the first observations of the xdat vector}

\item{h}{length of years covered by the information of the historical period}

\item{X0}{the perception threshold which is exceeded by the historical events. This is most likely different from the lowest historical value. It should indicate a value after which we are confident the event would have been recorded or left traces}

\item{binomialcens}{a logical value. Indicates whether the actual k values are to be used, or if only the information that the threshold X0 has been exceeded is used.}

\item{mulink}{the link function for the location parameter - default to identity}

\item{siglink}{the link function for the scale parameter - default to identity}

\item{shlink}{the link function for the shape parameter - default to identity}

\item{muinit}{initial values for the location parameter}

\item{siginit}{initial values for the scale parameter}

\item{shinit}{initial values for the shape parameter}

\item{show}{Logical; if \code{TRUE} (the default), print details of the fit.}

\item{method}{The optimization method (see \code{optim} for details)}

\item{maxit}{The maximum number of iterations}

\item{...}{Other control parameters for the optimization. These are passed to components of the control argument of optim.}
}
\value{
An object of the class \code{glo.fit}.
}
\description{
ML fitting procedures for sample of both systematic and historical data. 
Function structure inspired by \code{gev.fit} in ismev and similar to \code{\link{gev.hist.fit}}. 
Arguments naming and likelihood implementation follow Stedinger and Cohn (1986).
}
\examples{
set.seed(7821567)
xx <- rglo(500, 40, 6, -0.2)
xxsist <- xx[471:500]; xxhist <- xx[1:470][xx[1:470] > 80]
glo.hist.fit(c(xxhist,xxsist), k = length(xxhist), h = 470, X0 = 80)
glo.hist.fit(c(xxhist,xxsist), k = length(xxhist), h = 470, X0 = 80, binomialcens = TRUE)
glod.fit(xxsist) ## notice the higher standard errors
}
\references{
Hosking, J.R.M. and Wallis, J.R., 2005. Regional frequency analysis: an approach based on L-moments. Cambridge university press. 

Stedinger, J.R. and Cohn, T.A., 1986. Flood frequency analysis with historical and paleoflood information. Water resources research, 22(5), pp.785-793.

Macdonald, N., Kjeldsen, T.R., Prosdocimi, I. and Sangster, H., 2014. Reassessing flood frequency for the Sussex Ouse, Lewes: the inclusion of historical flood information since AD 1650. Natural Hazards and Earth System Sciences, 14(10), pp.2817-2828.
}
\seealso{
\code{\link{gev.hist.fit}}, \code{\link{dglo}}, Vignette on historical data
}
