% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gloq}
\alias{gloq}
\title{GLO quantile function}
\usage{
gloq(p, loc, scale, sh)
}
\arguments{
\item{loc,scale,sh}{the locatio, scale, shape parameters}

\item{p}{the non-exceedance probabiliy}
}
\description{
This function fits the GLo distribution. It's built mirroring the \code{gev.fit} function of \code{ismev}. This is the same as qglo, just used to mimic ismev
}

