% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsRFArevisited.R
\name{GOF tests}
\alias{GOF tests}
\alias{gofGEVtest_ip}
\alias{gofGLOtest_ip}
\title{Modified functions from nsRFA}
\usage{
gofGEVtest_ip(x, Nsim = 10000)

gofGLOtest_ip(x, Nsim = 10000)
}
\arguments{
\item{x}{the vector to be tested}

\item{Nsim}{the number of simulations which should be run}
}
\value{
The A2 statistics and the CORRECT p value
}
\description{
A modified version of the \code{nsRFA::gofXXXtest}. 
The main differences are in the pvalues: this one are correct, the original ones are 1-pvalue.
In the code lmoments are estimated using the lmom library
}
\concept{nsRFA GOF tests}
