% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gofGEVtest_ip}
\alias{gofGEVtest_ip}
\alias{gofGLOtest}
\title{Modified functions from nsRFA}
\usage{
gofGEVtest_ip(x, Nsim = 10000)
}
\arguments{
\item{x}{the vector to be tested}

\item{Nsim}{the number of simulations which should be run}
}
\value{
The A2 statistics and the CORRECT p value
}
\description{
Modified functions from nsRFA
}

