% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevglo.R
\name{Gumbel distribution}
\alias{Gumbel distribution}
\alias{dgum}
\alias{pgum}
\alias{rgum}
\alias{qggum}
\alias{qgum}
\title{The Gumbel distribution}
\usage{
dgum(x, loc, scale, log = FALSE)

pgum(q, loc, scale, lower.tail = TRUE, log.p = FALSE)

qgum(p, loc, scale, lower.tail = TRUE, log.p = FALSE)

rgum(n, loc, scale)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}}

\item{p}{vector of probabilities}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
dgum gives the density, pgum gives the distribution function, 
qgum gives the quantile function, and rgum generates random deviates.
The length of the result is determined by n for rgev, 
and is the maximum of the lengths of the numerical arguments for the other functions.
The numerical arguments are recycled to the length of the result. 
Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation for 
the Gumbel distribution with location parameter equal to \code{loc}, 
scale parameter equal to \code{scale}. This corresponds to a GEV 
distribution with shape parameter equal to 0.
}
\examples{
curve(dgum(x,4,6), from=-15, to = 40,type="l")
plot(ecdf(rgum(100,4,6)))
lines(seq(-15,40,by=0.5),pgum(seq(-15,40,by=0.5),4,6),col=2)
qgum(c(0.5,0.99,0.995,0.995,0.999),4,6) 
# notable quantiles
}
\concept{gumbel distribution}
\keyword{Gumbel}
\keyword{distribution}
