% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withHist.R
\name{lmom.hist.fit}
\alias{lmom.hist.fit}
\title{L-moment estimation in the presence of Historical Data}
\usage{
lmom.hist.fit(xdat, k = NULL, h = NULL, X0 = NULL, nmom = 3)
}
\arguments{
\item{xdat}{vector of historical and systematic/observed data - the first k elements of the vector should be the historical events}

\item{k}{number of historical events available. These events should be stored as the first observations of the xdat vector}

\item{h}{length of years covered by the information of the historical period}

\item{X0}{the perception threshold which is exceeded by the historical events. This is most likely different from the lowest historical value. It should indicate a value after which we are confident the event would have been recorded or left traces}

\item{nmom}{the number of moments to estimate}
}
\value{
A vector of L-moment, as lmom::samlmu.
}
\description{
L-moment fitting procedures for sample of both systematic and historical data via Partial Probability Weighted Moments (Wang, 1990).
}
\examples{
# library(ismev)
set.seed(54165784)
xx <- rgev(500, 40, 6, -0.2)
xxsist <- xx[471:500]; xxhist <- xx[1:470][xx[1:470] > 80]
lmh <- lmom.hist.fit(c(xxhist,xxsist), k = length(xxhist), h = 470, X0 = 80, nmom = 5)
lmh
lmom::pelgev(lmh)
gev.hist.fit(c(xxhist,xxsist), k = length(xxhist), h = 470, X0 = 80, show = FALSE)$mle
### note the different parametrization for the shape parameter
}
\references{
Wang, Q.J.(1990). Unbiased estimation of probability weighted moments and partial probability weighted moments from systematic and historical flood information and their application. Journal of hydrology, 120, 115--124.
}

