% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.gev.fit}
\alias{print.gev.fit}
\title{nicer print of gev.fit results}
\usage{
\method{print}{gev.fit}(fitobj)
}
\arguments{
\item{fitobj}{a fitted object of the class gev.fit}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\examples{
library(ismev)
print(gev.fit(c(50,45,65,78,12,23),show=FALSE))
}
\keyword{gev.fit}

