% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{print.glo.fit}
\alias{print.glo.fit}
\title{Nicer print of glo.fit results}
\usage{
\method{print}{glo.fit}(x, ...)
}
\arguments{
\item{x}{a fitted object of the class glo.fit.}

\item{...}{further arguments passed to \code{print}.}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\examples{
library(ismev)
print(glod.fit(c(50,45,65,78,12,23),show=FALSE))
}
\seealso{
glod.fit
}
