% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{print.gpd.fit}
\alias{print.gpd.fit}
\title{Nicer print of gpd.fit results}
\usage{
\method{print}{gpd.fit}(x, ...)
}
\arguments{
\item{x}{a fitted object of the class pp.fit.}

\item{...}{further arguments passed to \code{print}.}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\examples{
y <- c(53, 52, 49, 58, 50, 48, 47, 50, 46, 46, 49, 51, 47, 49, 50) 
a <- ismev::gpd.fit(y, threshold = 46, show=FALSE)
a
}
\keyword{gpd.fit}
