% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{print.kappa.fit}
\alias{print.kappa.fit}
\title{Nicer print of kappa.fit results}
\usage{
\method{print}{kappa.fit}(x, ...)
}
\arguments{
\item{x}{a fitted object of the class kappa.fit.}

\item{...}{further arguments passed to \code{print}.}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\seealso{
\code{\link{kappad.fit}}
}
