% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{print.pp.fit}
\alias{print.pp.fit}
\title{nicer print of pp.fit results}
\usage{
\method{print}{pp.fit}(obj)
}
\arguments{
\item{fitobj}{a fitted object of the class pp.fit}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\examples{
data(rain)
a <- ismev::pp.fit(rain, 10)
a
}
\keyword{pp.fit}

