% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesManip.R
\name{dates transformation}
\alias{date2month}
\alias{date2summer}
\alias{date2wy}
\alias{date2year}
\alias{dates transformation}
\title{Functions to transform dates into other useful information}
\usage{
date2year(datevec)

date2wy(datevec, endMonth = 9)

date2month(datevec)

date2summer(datevec, begMonth = 4, endMonth = 9)
}
\arguments{
\item{datevec}{a vetor of dates (\code{class(x)} must be Date)}
}
\value{
The year, water year or month from the input date. Or a 0-1 indicator of summer.
}

