% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Generalised Extreme Values distribution}
\alias{Generalised Extreme Values distribution}
\alias{dgev}
\alias{pgev}
\alias{qgev}
\alias{rgev}
\title{The Generalised Extreme Values distibution}
\usage{
dgev(x, loc, scale, sh)

pgev(q, loc, scale, sh, lower.tail = TRUE)

qgev(p, loc, scale, sh, lower.tail = TRUE)

rgev(n, loc, scale, sh)
}
\arguments{
\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{sh}{shape parameter}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{x,q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}}
}
\value{
These function mimic the standard output of distributions in R see \code{?pnorm}.
}
\description{
Density, distribution function, quantile function and random generation for the Generalised Extreme Values distribution with location parameter equal to \code{loc}, scale paraemter equale to \code{scale} and shape parameter equal to \code{shape}
}
\details{
Extra care should be taken on the shape parameter of the distribution as the HW and the Coles' version differ in the sign of the shape parameter: be careful!
}
\examples{
plot(seq(-15,40,by=0.2),dgev(seq(-15,40,by=0.2),4,6,0.2),type="l")
plot(ecdf(rgev(100,4,6,0.2)))
lines(seq(-15,40,by=0.5),pgev(seq(-15,40,by=0.5),4,6,0.2),col=2)
qgev(c(0.5,0.99,0.995,0.995,0.999),4,6,0.2) # the 1-in-2years, 1-in-100years, 1-in-200years, 1-in-1000years events
}
\keyword{Extreme}
\keyword{Generalised}
\keyword{Values}
\keyword{distibution}

