% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{glo.fit}
\alias{glo.fit}
\title{Maximum Likelihood Fitting for the GLO distibution}
\usage{
glo.fit(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, ...)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted}

\item{ydat}{A matrix of covariates for generalized linear modelling of the parameters (or NULL (the default) for stationary fitting). The number of rows should be the same as the length of xdat}

\item{mul}{Numeric vectors of integers, giving the columns of ydat that contain covariates for generalized linear modelling of the location parameter (or NULL (the default) if the corresponding parameter is stationary)}

\item{sigl}{As \code{mul} for the scale parameter}

\item{shl}{As \code{mul} for the shape parameter}

\item{mulink}{the link function for the location parameter - default to identity}

\item{siglink}{the link function for the scale parameter - default to identity}

\item{shlink}{the link function for the shape parameter - default to identity}

\item{muinit}{initial values for the location parameter}

\item{siginit}{initial values for the scale parameter}

\item{shinit}{initial values for the shape parameter}

\item{method}{The optimization method (see \code{optim} for details)}

\item{maxit}{The maximum number of iterations}

\item{...}{Other control parameters for the optimization. These are passed to components of the control argument of optim.}
}
\value{
an object of the glo.fit class - with values which mirror the ones of the gev.fit class
}
\description{
This function has the same structure as the \code{gev.fit} from \code{ismev}
}
\examples{
set.seed(5846)
print(glo.fit(rglo(n = 80, 50, 6, -0.2), show=FALSE))
}
\keyword{glo.fit}

