% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismevExtension.R
\name{gloq}
\alias{gloq}
\title{GLO quantile function}
\usage{
gloq(p, loc, scale, sh)
}
\arguments{
\item{p}{the non-exceedance probabiliy}

\item{loc, scale, sh}{the locatio, scale, shape parameters}
}
\description{
This function gives quantiles the GLo distribution. It's built mirroring the \code{gev.fit} function of \code{ismev}. This is the same as qglo, just used to mimic ismev
}

