% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lengthNOTna}
\alias{lengthNOTna}
\title{Add together two numbers}
\usage{
lengthNOTna(x)
}
\arguments{
\item{x}{a vector}
}
\value{
The length of \code{x} without the NA
}
\description{
Add together two numbers
}
\examples{
x <- c(5,12,NA,4)
lengthNOTna(x)
}

