% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lengthNOTna}
\alias{lengthNOTna}
\title{Length without NAs}
\usage{
lengthNOTna(x)
}
\arguments{
\item{x}{a vector}
}
\value{
The length of the vector when NAs are excluded
}
\description{
Length without NAs
}
\examples{
x <- c(5,12,NA,4)
lengthNOTna(x)
}

