% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.glo.fit}
\alias{print.glo.fit}
\title{nicer print of glo.fit results}
\usage{
\method{print}{glo.fit}(fitobj)
}
\arguments{
\item{fitobj}{a fitted object of the class glo.fit}
}
\description{
This function prints the MLE, se, convergence info and negative log-likelihood value.
}
\examples{
library(ismev)
print(glo.fit(c(50,45,65,78,12,23),show=FALSE))
}
\keyword{glo.fit}

